requirements.txt
Quart
discord
discord.py
discord.ui
jishaku
asyncio
discord-ext-menus
datetime
flask
typing
psutil==5.9.8
collection
requests
pathlib
pymongo
wavelink>=3.4.1
httpx
discord.ext-context
tasksio
motor
Augmentor
numpy
aiosqlite
lyricsgenius
gtts
langdetect
python-dotenv
pyttsx3
PyYAML
Requests
openai
duckduckgo-search==6.3.6
pillow==10.0.1
colorama
chess
python-dateutil
topggpy
yt-dlp
spotipy
lyricsgenius
groq
pymongo[srv]
tenacity
sympy
numpy
lyricsgenius
matplotlib
google-generativeai
google-generativeai
aiohttp
aiosqlite
chess
colorama
discord.py
duckduckgo-search
flask
google-generativeai
groq
gtts
jishaku
langdetect
lyricsgenius
matplotlib
numpy
openai
psutil
pymongo
python-dateutil
python-dotenv
pyttsx3
PyYAML
requests
spotipy
sympy
tenacity
wavelink
yt-dlp
eventlet
Flask
Flask-SocketIO
python-decouple
requests
google-genai
aiofiles
aiohttp
google-genai
pillow
google-generativeai
aiosqlite
colorama
discord.py
jishaku
python-dotenv